# backend/api-fastapi/app/models/schemas.py
from pydantic import BaseModel

class UserCreate(BaseModel):
    name: str

class UserRead(BaseModel):
    id: int
    name: str

    class Config:
        from_attributes = True  # substitui orm_mode